[tt-llk] Consolidate Quasar unary-SFPU tests into one python + one cpp
Replace the four standalone Quasar unary-SFPU test pairs (abs, square,
rsqrt, nonlinear) with a single python driver + single cpp source,
mirroring how Blackhole drives eltwise_unary_sfpu_test.cpp. The op is
selected at compile time via SFPU_UNARY_OPERATION = SfpuType::<op> and
dispatched through a new Quasar-specific helper.

- Add tests/helpers/include/sfpu_operations_quasar.h: maps SfpuType to the
  Quasar _calculate_*_/_init_*_ entry points for all 11 ops (the Quasar
  analog of Blackhole's sfpu_operations.h, whose metal SFPU_CALL macros and
  signatures do not apply to Quasar).
- Add sources/quasar/eltwise_unary_sfpu_quasar_test.cpp: keeps the existing
  unpack_to_dest dual-path scheme verbatim; only the SFPU op call is now
  dispatched.
- Add python_tests/quasar/test_eltwise_unary_sfpu_quasar.py: one driver with
  a per-op config table that preserves each op's exact original sweep.
- Delete the abs/square/rsqrt/nonlinear cpp + python pairs.
- Make test_sfpu_square_trisc3_quasar.py self-contained (it imported from the
  now-deleted test_sfpu_square_quasar.py).

Parity: consolidated collects 1170 variants = abs(78) + square(156) +
rsqrt(104) + nonlinear(832), matching the four standalone files. Simulator
smoke test passes 4/4 across both unpack-to-dest and FPU-datacopy paths.
