+++ tt_metal/tt-llk/tests/helpers/include/sfpu_operations_quasar.h
@@ -1,0 +1,128 @@
+// SPDX-FileCopyrightText: © 2026 Tenstorrent AI ULC
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include <cstdint>
+
+#include "ckernel.h"
+#include "llk_defs.h"
+#include "llk_math_eltwise_unary_sfpu.h"
+
+// Quasar unary-SFPU operation headers. Quasar exposes lower-level
+// `_calculate_*_` / `_init_*_` entry points (not the metal `llk_sfpu/`
+// wrappers that the Blackhole `sfpu_operations.h` dispatches through), so this
+// header is the Quasar-specific equivalent of that dispatcher.
+//
+// To add a new Quasar unary SFPU operation:
+// 1. Include its `ckernel_sfpu_<op>.h` below.
+// 2. Add the `SfpuType` enumerator to the `if constexpr` chain in
+//    call_unary_sfpu_operation_quasar() (and to init_unary_sfpu_operation_quasar()
+//    if the op needs an init step).
+#include "experimental/ckernel_sfpu_abs.h"
+#include "sfpu/ckernel_sfpu_exp.h"
+#include "sfpu/ckernel_sfpu_gelu.h"
+#include "sfpu/ckernel_sfpu_recip.h"
+#include "sfpu/ckernel_sfpu_relu.h"
+#include "sfpu/ckernel_sfpu_rsqrt.h"
+#include "sfpu/ckernel_sfpu_sigmoid.h"
+#include "sfpu/ckernel_sfpu_silu.h"
+#include "sfpu/ckernel_sfpu_sqrt.h"
+#include "sfpu/ckernel_sfpu_square.h"
+#include "sfpu/ckernel_sfpu_tanh.h"
+
+namespace test_utils
+{
+using namespace ckernel;
+using namespace ckernel::sfpu;
+
+/**
+ * @brief Run the per-operation init step for a Quasar unary SFPU op.
+ *
+ * Most Quasar unary ops need no dedicated init (the shared
+ * `_llk_math_eltwise_sfpu_init_()` in the kernel is sufficient); only the ops
+ * with a `_init_*_` entry point are handled here. This mirrors the inline
+ * dispatcher previously embedded in `sfpu_nonlinear_quasar_test.cpp`, which
+ * initialised only gelu.
+ *
+ * @tparam OPERATION The SFPU operation type (compile-time `SfpuType` constant).
+ * @note Pair with @ref call_unary_sfpu_operation_quasar for the calculate step.
+ */
+template <SfpuType OPERATION>
+void init_unary_sfpu_operation_quasar()
+{
+    if constexpr (OPERATION == SfpuType::gelu)
+    {
+        _init_gelu_();
+    }
+}
+
+/**
+ * @brief Apply a Quasar unary SFPU op in-place on one Dest tile.
+ *
+ * Selects the matching `_calculate_*_` entry point at compile time from the
+ * kernel's `SFPU_UNARY_OPERATION` constant and forwards it to
+ * `_llk_math_eltwise_unary_sfpu_params_`. `ITERATIONS` defaults to
+ * `SFPU_ITERATIONS` (from `llk_defs.h`); abs/rsqrt are passed it explicitly so
+ * they override their own per-op default of 8, matching the standalone tests.
+ *
+ * @tparam OPERATION The SFPU operation type (compile-time `SfpuType` constant).
+ * @tparam ITERATIONS Number of SFPU loop iterations.
+ * @param dst_index Destination tile index operated on (already offset by DST_INDEX).
+ * @note Must be preceded by @ref init_unary_sfpu_operation_quasar for the same op.
+ */
+template <SfpuType OPERATION, int ITERATIONS = SFPU_ITERATIONS>
+void call_unary_sfpu_operation_quasar(std::uint32_t dst_index)
+{
+    if constexpr (OPERATION == SfpuType::abs)
+    {
+        _llk_math_eltwise_unary_sfpu_params_(_calculate_abs_<ITERATIONS>, dst_index);
+    }
+    else if constexpr (OPERATION == SfpuType::exponential)
+    {
+        _llk_math_eltwise_unary_sfpu_params_(_calculate_exp_<true /* APPROX */, ITERATIONS>, dst_index);
+    }
+    else if constexpr (OPERATION == SfpuType::gelu)
+    {
+        _llk_math_eltwise_unary_sfpu_params_(_calculate_gelu_<ITERATIONS>, dst_index);
+    }
+    else if constexpr (OPERATION == SfpuType::relu)
+    {
+        _llk_math_eltwise_unary_sfpu_params_(_calculate_relu_<ITERATIONS>, dst_index);
+    }
+    else if constexpr (OPERATION == SfpuType::reciprocal)
+    {
+        _llk_math_eltwise_unary_sfpu_params_(_calculate_reciprocal_<true /* APPROX */, ITERATIONS>, dst_index);
+    }
+    else if constexpr (OPERATION == SfpuType::sqrt)
+    {
+        _llk_math_eltwise_unary_sfpu_params_(_calculate_sqrt_<true /* APPROX */, ITERATIONS>, dst_index);
+    }
+    else if constexpr (OPERATION == SfpuType::tanh)
+    {
+        _llk_math_eltwise_unary_sfpu_params_(_calculate_tanh_<true /* APPROX */, ITERATIONS>, dst_index);
+    }
+    else if constexpr (OPERATION == SfpuType::sigmoid)
+    {
+        _llk_math_eltwise_unary_sfpu_params_(_calculate_sigmoid_<ITERATIONS>, dst_index);
+    }
+    else if constexpr (OPERATION == SfpuType::silu)
+    {
+        _llk_math_eltwise_unary_sfpu_params_(_calculate_silu_<ITERATIONS>, dst_index);
+    }
+    else if constexpr (OPERATION == SfpuType::rsqrt)
+    {
+        _llk_math_eltwise_unary_sfpu_params_(_calculate_rsqrt_<ITERATIONS>, dst_index);
+    }
+    else if constexpr (OPERATION == SfpuType::square)
+    {
+        _llk_math_eltwise_unary_sfpu_params_(_calculate_square_<ITERATIONS>, dst_index);
+    }
+    else
+    {
+        LLK_ASSERT(false, "Unsupported Quasar unary SFPU operation");
+    }
+}
+
+} // namespace test_utils
