--- tt_metal/tt-llk/tests/python_tests/quasar/test_sfpu_square_quasar.py
@@ -1,269 +1,0 @@
-# SPDX-FileCopyrightText: © 2026 Tenstorrent AI ULC
-# SPDX-License-Identifier: Apache-2.0
-
-import math
-from typing import List
-
-import pytest
-import torch
-from helpers.format_config import DataFormat, FormatConfig
-from helpers.golden_generators import UnarySFPUGolden, get_golden_generator
-from helpers.llk_params import (
-    DataCopyType,
-    DestAccumulation,
-    DestSync,
-    ImpliedMathFormat,
-    MathOperation,
-    UnpackerEngine,
-    format_dict,
-)
-from helpers.param_config import (
-    input_output_formats,
-    is_invalid_quasar_sfpu_format_combination,
-    parametrize,
-)
-from helpers.stimuli_config import StimuliConfig
-from helpers.stimuli_generator import generate_stimuli
-from helpers.test_config import TestConfig
-from helpers.test_variant_parameters import (
-    DATA_COPY_TYPE,
-    DEST_INDEX,
-    DEST_SYNC,
-    IMPLIED_MATH_FORMAT,
-    MATH_OP,
-    NUM_FACES,
-    TEST_FACE_DIMS,
-    TILE_COUNT,
-    UNPACKER_ENGINE_SEL,
-)
-from helpers.utils import passed_test
-
-
-def prepare_square_inputs(
-    src_A: torch.Tensor,
-    src_B: torch.Tensor,
-    input_format: DataFormat,
-    output_format: DataFormat,
-) -> torch.Tensor:
-    """
-    Prepare input tensor for square operation with safe value ranges.
-
-    Applies log-uniform distribution and clamps values to ensure:
-    - Input values fit in the input format
-    - Squared output values fit in the output format
-    - Good test coverage across orders of magnitude
-
-    Args:
-        src_A: Source tensor A (used for magnitude distribution)
-        src_B: Source tensor B (used for sign distribution)
-        input_format: Input data format
-        output_format: Output data format
-
-    Returns:
-        Prepared tensor with safe values for squaring
-    """
-    input_torch_format = format_dict[input_format]
-    output_torch_format = format_dict[output_format]
-    input_finfo = torch.finfo(input_torch_format)
-    output_finfo = torch.finfo(output_torch_format)
-
-    # For squaring, x² must fit in the OUTPUT format
-    max_safe_value = math.sqrt(output_finfo.max) * 0.9
-
-    # Special handling for bfloat16: it has wide range but limited precision
-    # Extreme values lose precision, so limit to reasonable bounds
-    if input_torch_format == torch.bfloat16:
-        # Limit to range where squaring maintains reasonable precision
-        max_safe_value = min(max_safe_value, 1e4)  # 10000² = 1e8 fits comfortably
-    else:
-        # For Float16, ensure input itself fits in input format
-        # Float16 max is ~65504, so sqrt(65504) ≈ 256 is the safe limit
-        max_safe_value = min(max_safe_value, math.sqrt(input_finfo.max) * 0.9)
-
-    min_magnitude = max(1e-6, input_finfo.tiny * 100)  # Avoid denormals
-
-    # Ensure src_A and src_B don't contain inf/nan before normalization
-    src_A_float = src_A.to(torch.float32)
-    src_B_float = src_B.to(torch.float32)
-
-    # Normalize src_A to [0, 1] range for log-uniform distribution
-    src_A_min = src_A_float.min()
-    src_A_max = src_A_float.max()
-    src_A_normalized = (
-        (src_A_float - src_A_min) / (src_A_max - src_A_min)
-        if src_A_max > src_A_min
-        else torch.zeros_like(src_A_float)
-    )
-
-    # Use log-uniform distribution for magnitudes to test across orders of magnitude
-    log_min = torch.log(torch.tensor(min_magnitude, dtype=torch.float32))
-    log_max = torch.log(torch.tensor(max_safe_value, dtype=torch.float32))
-    magnitudes = torch.exp(log_min + src_A_normalized * (log_max - log_min))
-
-    # Randomly assign signs to get both positive and negative values
-    src_B_min = src_B_float.min()
-    src_B_max = src_B_float.max()
-    src_B_normalized = (
-        (src_B_float - src_B_min) / (src_B_max - src_B_min)
-        if src_B_max > src_B_min
-        else torch.zeros_like(src_B_float)
-    )
-    signs = torch.where(src_B_normalized < 0.5, -1.0, 1.0)
-
-    # Apply signs and clamp to safe range BEFORE converting to input format
-    src_A_values = signs * magnitudes
-    src_A_values = torch.clamp(src_A_values, -max_safe_value, max_safe_value)
-    result = src_A_values.to(input_torch_format)
-
-    return result
-
-
-def generate_sfpu_square_combinations(
-    formats_list: List[FormatConfig],
-):
-    """
-    Generate SFPU square test combinations.
-
-    Args: Input-output format pairs
-
-    Returns: List of (format, dest_acc, dest_sync, implied_math_format, input_dimensions) tuples
-    """
-    combinations = []
-
-    dest_sync_modes = (DestSync.Half, DestSync.Full)
-    for fmt in formats_list:
-        in_fmt = fmt.input_format
-
-        dest_acc_modes = (
-            (DestAccumulation.Yes,)
-            if in_fmt.is_32_bit()
-            else (DestAccumulation.No, DestAccumulation.Yes)
-        )
-        for dest_acc in dest_acc_modes:
-            # Skip invalid format combinations for Quasar
-            if is_invalid_quasar_sfpu_format_combination(fmt, dest_acc):
-                continue
-
-            for dest_sync in dest_sync_modes:
-                for implied_math_format in [
-                    ImpliedMathFormat.No,
-                    ImpliedMathFormat.Yes,
-                ]:
-                    for input_dimensions in [[32, 32], [64, 64], [32, 64]]:
-                        combinations.append(
-                            (
-                                fmt,
-                                dest_acc,
-                                dest_sync,
-                                implied_math_format,
-                                input_dimensions,
-                            )
-                        )
-
-    return combinations
-
-
-SFPU_SQUARE_FORMATS = input_output_formats(
-    [
-        DataFormat.Float16,
-        DataFormat.Float32,
-        DataFormat.Float16_b,
-    ]
-)
-
-
-@pytest.mark.quasar
-@parametrize(
-    formats_dest_acc_sync_implied_math_input_dims=generate_sfpu_square_combinations(
-        SFPU_SQUARE_FORMATS
-    ),
-)
-def test_sfpu_square_quasar(formats_dest_acc_sync_implied_math_input_dims):
-    """
-    Test square operation on Quasar architecture.
-
-    Uses PyTorch's square (x**2) as the golden reference and generates input stimuli
-    covering the full representable range
-    """
-    (formats, dest_acc, dest_sync, implied_math_format, input_dimensions) = (
-        formats_dest_acc_sync_implied_math_input_dims[0]
-    )
-
-    # Set seed for reproducibility
-    torch.manual_seed(42)
-
-    src_A, tile_cnt_A, src_B, _ = generate_stimuli(
-        stimuli_format_A=formats.input_format,
-        input_dimensions_A=input_dimensions,
-        stimuli_format_B=formats.input_format,
-        input_dimensions_B=input_dimensions,
-    )
-
-    # Prepare inputs with safe ranges for squaring operation
-    src_A = prepare_square_inputs(
-        src_A, src_B, formats.input_format, formats.output_format
-    )
-
-    num_faces = 4
-
-    generate_golden = get_golden_generator(UnarySFPUGolden)
-    golden_tensor = generate_golden(
-        MathOperation.Square,
-        src_A,
-        formats.output_format,
-        dest_acc,
-        formats.input_format,
-        input_dimensions,
-    )
-
-    unpack_to_dest = (
-        formats.input_format.is_32_bit() and dest_acc == DestAccumulation.Yes
-    )
-    configuration = TestConfig(
-        "sources/quasar/sfpu_square_quasar_test.cpp",
-        formats,
-        templates=[
-            MATH_OP(mathop=MathOperation.Square),
-            IMPLIED_MATH_FORMAT(implied_math_format),
-            DATA_COPY_TYPE(DataCopyType.A2D),
-            UNPACKER_ENGINE_SEL(
-                UnpackerEngine.UnpDest if unpack_to_dest else UnpackerEngine.UnpA
-            ),
-            DEST_SYNC(dest_sync),
-        ],
-        runtimes=[
-            TILE_COUNT(tile_cnt_A),
-            NUM_FACES(num_faces),
-            TEST_FACE_DIMS(),
-            DEST_INDEX(0),
-        ],
-        variant_stimuli=StimuliConfig(
-            src_A,
-            formats.input_format,
-            src_B,
-            formats.input_format,
-            formats.output_format,
-            tile_count_A=tile_cnt_A,
-            tile_count_B=tile_cnt_A,
-            tile_count_res=tile_cnt_A,
-            num_faces=num_faces,
-        ),
-        unpack_to_dest=(
-            formats.input_format.is_32_bit() and dest_acc == DestAccumulation.Yes
-        ),
-        dest_acc=dest_acc,
-    )
-
-    res_from_L1 = configuration.run().result
-
-    # Verify results match golden
-    assert len(res_from_L1) == len(
-        golden_tensor
-    ), "Result tensor and golden tensor are not of the same length"
-
-    torch_format = format_dict[formats.output_format]
-    res_tensor = torch.tensor(res_from_L1, dtype=torch_format)
-
-    assert passed_test(
-        golden_tensor, res_tensor, formats.output_format
-    ), "Assert against golden failed"
